WELD-1791 Revisit configuration
- introduce WeldConfiguration per-deployment service
- supported configuration keys can be found in ConfigurationKey enum

+++ impl/src/main/java/org/jboss/weld/bean/proxy/ProxyFactory.java
@@ -50,7 +50,7 @@ import org.jboss.classfilewriter.code.CodeAttribute;
 import org.jboss.classfilewriter.util.Boxing;
 import org.jboss.classfilewriter.util.DescriptorUtils;
 import org.jboss.weld.Container;
-import org.jboss.weld.SystemPropertiesConfiguration;
+import org.jboss.weld.config.WeldConfiguration;
 import org.jboss.weld.exceptions.DefinitionException;
 import org.jboss.weld.exceptions.WeldException;
 import org.jboss.weld.interceptor.proxy.LifecycleMixin;
@@ -97,6 +97,8 @@ public class ProxyFactory<T> {
     private final Bean<?> bean;
     private final Class<?> proxiedBeanType;
     private final String contextId;
+
+    private final WeldConfiguration configuration;
 
     public static final String CONSTRUCTED_FLAG_NAME = "constructed";
 
@@ -155,6 +157,7 @@ public class ProxyFactory<T> {
         this.bean = bean;
         this.contextId = contextId;
         this.proxiedBeanType = proxiedBeanType;
+        this.configuration = Container.instance(contextId).deploymentManager().getServices().get(WeldConfiguration.class);
         addInterfacesFromTypeClosure(typeClosure, proxiedBeanType);
         TypeInfo typeInfo = TypeInfo.of(typeClosure);
         Class<?> superClass = typeInfo.getSuperClass();
@@ -458,9 +461,8 @@ public class ProxyFactory<T> {
         // TODO: change the ProxyServices SPI to allow the container to figure out
         // which PD to use
 
-        if (SystemPropertiesConfiguration.INSTANCE.isProxyDumpEnabled()) {
-            dumpToFile(SystemPropertiesConfiguration.INSTANCE.getProxyDumpPath(), proxyClassName, proxyClassType.toBytecode());
-        }
+        // Dump proxy type bytecode if necessary
+        dumpToFile(proxyClassName, proxyClassType.toBytecode());
 
         ProtectionDomain domain = AccessController.doPrivileged(new GetProtectionDomainAction(proxiedBeanType));
 
@@ -475,9 +477,13 @@ public class ProxyFactory<T> {
         return proxyClass;
     }
 
-    private void dumpToFile(File dumpPath, String fileName, byte[] data) {
+    private void dumpToFile(String fileName, byte[] data) {
+        File proxyDumpFilePath = configuration.getProxyDumpFilePath();
+        if(proxyDumpFilePath == null) {
+            return;
+        }
         FileOutputStream dumpFileStream = null;
-        File classFile = new File(dumpPath, fileName + ".class");
+        File classFile = new File(proxyDumpFilePath, fileName + ".class");
         try {
             if (classFile.isFile()) {
                 if (!classFile.delete()) {
